MNT: Rename some subpackages.
This denotes them as internal APIs.
