+++ metpy/io/_tools.py
